Document how to override certain settings
This commit mainly just adds a bunch of documentation to the `settings`
module and README outlining how to go about overriding certain settings, should
people need to do that. It also removes DEFAULT_UNIT_OPTIONS, a dict
that was previously passed to the functions in `factories.py`. I removed
this because it was unnecessary and would otherwise serve to override
your overrides with the original default values.