Try getting artifact upload/download to work
Tox isn't recognizing the wheel file in the dist directory after
downloading the artifact. From the examples it looks like the download
action should drop the artifact into the working directory, but since
the artifact *is* a directory it may not be working the way I think it
should.

I'm trying to see if tarring/untarring the dist directory will help.