Reinstate py35 in GH actions wf, on ubuntu 20.04
This is why I shouldn't try to do quick changes before leaving for the
day. Looks like GH runners are changing over `ubuntu-latest` from 20.04
to 22.04, where 22.04 doesn't support Python 3.5 or 3.6. But there's no
need to remove the py3.5 or 3.6 tests if we specify we need 20.04 in the
GH actions workflow.